Add blfabrstmplt-s03.json and modify README.md (#3)
* Create blfabrstmplt-s03.json

This will setup a RHEL 7.8 that uses LVM for the OS disk and it has two additional disks available for lab exercises.
Custom data has been updated to enable firewall and SELinux support as these are enable on RHEL 7

* Create blfabrstmplt-s03.sh

Temporally while we fix the customdata

* Update blfabrstmplt-s03.sh

* Update blfabrstmplt-s03.sh

* Update blfabrstmplt-s03.sh

Force to use firewall-offline-cmd and not firewall-cmd

* Update blfabrstmplt-s03.sh

* Update blfabrstmplt-s03.sh

* Update blfabrstmplt-s03.sh

* Update blfabrstmplt-s03.sh

* Update blfabrstmplt-s03.sh

* Update blfabrstmplt-s03.sh

* Remove --permanent from firewall-offline-cmd

Remove --permanent from firewall-offline-cmd

* Change Custom Data content

After fixing bug with firewalld and cloud-init. Testing final content on the custom data for cloud-init

* Final fix for custom data

A sleep is required betwen the stop the add and the start on firewalld for some reason

* Create blfabrstmplt-s03.md

Initial readme for session 3

* Update README.md

Add URLs

* Update README.md

* Update README.md

* Update README.md

* Delete blfabrstmplt-s03.sh

This file is not required anymore as the issue with cloud-init has been fixed for blfabrstmplt-s03.json

* Delete blfabrstmplt-s03.md

No need as the link of the json template has been added to the main README.md

* Add storage account for diagnostic profile

Adding the storage account for the diagnostic profile, the console of the VM should work out of the box

* Replacing with the correct json file

Incorrect uploaded json

* Update README.md

Reworded VM as Scenario. Multiple resources are being created, not just a VM.
Future scenarios may contain multiple VMs, too.
We may want to add further instructions or requirements. Maybe even a description of each scenario.

* Changing from RHEL 7.8 to 7.9

As per suggestion. Changing to the minor latest version available for RHEL 7

Co-authored-by: adelgadohell <alf.e@microsoft.com>